distributed post box component into 5 components then colled them all in the post .jsx file
